feat: variable shadowing support, signed type(intN).min/max, int64 array sizes
Multiple improvements accumulated from prior work sessions:

- Variable shadowing: ExpressionBuilder tracks var name→declaration ID mappings
  with ScopeGuard save/restore. Inner-scope declarations that shadow outer ones
  get unique names (name__id). SolIdentifier and SolVariableDeclaration use
  resolveVarName/lookupVarName for consistent naming.

- Signed type(intN).min/max: SolMetaTypeAccess now computes correct values for
  signed integer types (e.g. type(int8).min = 128 as two's complement biguint,
  type(int8).max = 127).

- int → int64_t for array sizes: WType, TypeMapper, TypeCoercion, AssemblyBuilder,
  and ContractSplitter all use int64_t for array sizes to support large arrays.

- AWSTBuilder: register named return variable names in library/free function
  translation so inner scoping detects shadowing correctly.

+++ src/awst/WType.h
@@ -179,7 +179,7 @@ private:
 class ARC4StaticArray: public WType
 {
 public:
-	ARC4StaticArray(WType const* _elementType, int _arraySize)
+	ARC4StaticArray(WType const* _elementType, int64_t _arraySize)
 		: WType(
 			  "arc4.static_array<" + _elementType->name() + ", "
 				  + std::to_string(_arraySize) + ">",
@@ -193,11 +193,11 @@ public:
 
 	std::string jsonType() const override { return "ARC4StaticArray"; }
 	WType const* elementType() const { return m_elementType; }
-	int arraySize() const { return m_arraySize; }
+	int64_t arraySize() const { return m_arraySize; }
 
 private:
 	WType const* m_elementType;
-	int m_arraySize;
+	int64_t m_arraySize;
 };
 
 class ARC4Struct: public WType
@@ -231,7 +231,7 @@ public:
 	explicit ReferenceArray(
 		WType const* _elementType,
 		bool _immutable = true,
-		std::optional<int> _arraySize = std::nullopt
+		std::optional<int64_t> _arraySize = std::nullopt
 	)
 		: WType(
 			  _arraySize
@@ -247,11 +247,11 @@ public:
 
 	std::string jsonType() const override { return "ReferenceArray"; }
 	WType const* elementType() const { return m_elementType; }
-	std::optional<int> arraySize() const { return m_arraySize; }
+	std::optional<int64_t> arraySize() const { return m_arraySize; }
 
 private:
 	WType const* m_elementType;
-	std::optional<int> m_arraySize;
+	std::optional<int64_t> m_arraySize;
 };
 
 class WTuple: public WType
